Add files via upload
For #3, I created a function called regularize(a) that takes an argument a={0, .5 ,1}. It does all of #3, for the Hitters dataset (excluding random forest). The only difference between between ridge, elnet, and lasso is the value of alpha given in this formula: <img src="https://i.imgur.com/ncWbRGG.png">. Thus, supplying a=0 completes #3 for ridge, a=.5 for elnet, and a=1 for lasso. It outputs the train and test R^2 values according to the value of a given. 

For #4(c), a second argument regularize(a,plot=boolean) can be supplied to additionally output a 10 Fold CV plot. The default value is plot=FALSE.

The regularize() function itself is long, so I use source() to call upon it in the main body, which I have called "Working Draft". Of course, set the working directory to wherever "Regularization Function.R" to call the function without error.  We can then create a dataframe of R^2's using rbind and cbind as illustrated. From this dataframe, we can produce the boxplots for #4(b).